services/order-service/requirements.txt
flask==3.1.1
werkzeug==3.1.3
requests==2.32.3
prometheus-client==0.21.1
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation-flask==0.49b2
opentelemetry-instrumentation-requests==0.49b2
# flask==3.1.0
# prometheus-client==0.21.1
# opentelemetry-api==1.29.0
# opentelemetry-sdk==1.29.0
# opentelemetry-exporter-otlp-proto-grpc==1.29.0
# opentelemetry-instrumentation-flask==0.50b0
# opentelemetry-instrumentation-requests==0.50b0
# requests==2.32.3
